+++ src/serlio/materials/StingrayMaterialNode.h
